1.将一些在一个页面以上用到的请求挪到common.js中。2.侧边栏
