+++ src/components/common/Sidebar.vue
@@ -2,8 +2,8 @@
 <template>
 	<div class="sidebar">
 		<div class="sidebar-btn">
-			<el-button type="warning" icon="el-icon-star-off" circle></el-button>
-			<el-button type="success" icon="el-icon-arrow-up" circle></el-button>
+			<el-button type="warning" icon="el-icon-star-off" circle @click="myConcern"></el-button>
+			<el-button type="danger" icon="el-icon-arrow-up" circle></el-button>
 		</div>
 		
 		<div class="sidebar-spread">
@@ -26,9 +26,28 @@
     
     computed: {},
     
-    methods: {}
+    methods: {
+      myConcern(){
+      
+      }
+    }
   }
 </script>
 
 <style scoped lang="scss">
+	.sidebar{
+		position: fixed;
+		bottom: 20px;
+		right: 20px;
+		z-index: 100;
+	}
+	
+	.sidebar-btn{
+		width: 40px;
+	}
+	
+	.el-button+.el-button{
+		margin-left: 0;
+		margin-top: 10px;
+	}
 </style>
